4415: Remove unused code: SicCode.Created: Related code changes
