+++ GenderPayGap.Database/Models/SicCode.cs
@@ -6,7 +6,7 @@ using Newtonsoft.Json;
 namespace GenderPayGap.Database
 {
     [JsonObject(MemberSerialization.OptIn)]
-    public partial class SicCode
+    public class SicCode
     {
 
         public SicCode()
@@ -27,6 +27,18 @@ namespace GenderPayGap.Database
 
         public virtual SicSection SicSection { get; set; }
         public virtual ICollection<OrganisationSicCode> OrganisationSicCodes { get; set; }
+
+        public override bool Equals(object obj)
+        {
+            // Check for null values and compare run-time types.
+            if (obj == null || GetType() != obj.GetType())
+            {
+                return false;
+            }
+
+            var target = (SicCode) obj;
+            return SicCodeId == target.SicCodeId;
+        }
 
     }
 }
